Verify Previously Saved Cluster Data

Users may wish to save long-form combined cluster data as a .csv file or similar after combining individual clusters with qm_combine. Unless data are saved as an .rda file, the class definition of "qm_cluster" will be lost. The qm_verify function allows users to import data from any file type readable by R, verify that it has the column names needed for qm_summarize, and re-add the "qm_cluster" class.

qm_verify(clusters)

Arguments

clusters
An object created by qm_combine with two or more clusters worth of data that has been previously saved and requires verification before summarization.

Value

A tibble stored with a custom class of qm_cluster to facilitate data validation.